P3: define kernel text ctype byte domains
Finding: CLUSTER-ctype-kernel-text-plain-char

Kernel pseudo-file, proc symlink, process command-line/name, and collector identifier bytes were passed from plain char to isdigit(), iscntrl(), and tolower(). On signed-char targets, high-bit bytes therefore promoted to negative values outside the unsigned char/EOF domain and invoked undefined behavior.

Convert each of eight ctype arguments to uint8_t at the call boundary. This covers debugfs extfrag chart IDs, /proc/interrupts parsing, mountinfo escape decoding, cpuidle dimension IDs, local-socket cmdline and inode-link parsing, and cgroup comm classification. Seven sites are current-capable; the hardirq parser is potential-only because its sole caller is commented out.

ASCII and existing parser grammar, return values, output, allocation, lifetime, locking, and ownership remain unchanged. High-bit bytes now use the existing C-locale classification. Each fix is constant-time with no new branch, scan, I/O, allocation, or lock.

Validated 54 GCC/Clang production and consumer object builds across default, signed-char, and unsigned-char modes. The 256-byte probes passed with both compilers, including ASan and UBSan. The final complete six-reviewer gate approved the unchanged target, and the full raw process audit passed.

(cherry picked from commit 14eab79f0ef407c79b766bd7f7d4f8e4d5035361)

+++ src/collectors/cgroups.plugin/cgroup-discovery.c
@@ -884,7 +884,7 @@ static inline void discovery_find_all_cgroups_v2() {
 
 static int is_digits_only(const char *s) {
     do {
-        if (!isdigit(*s++)) {
+        if (!isdigit((uint8_t)*s++)) {
             return 0;
         }
     } while (*s);
